+++ BrasserieOrdersysteem/Models/OrderRule.cs
@@ -3,8 +3,8 @@
     public class OrderRule
     {
         public int Id { get; set; }
-        public double Units { get; set; }
-        public double PricePerUnit { get; set; }
+        public int Units { get; set; }
+        public decimal PricePerUnit { get; set; }
 
         public int OrderId { get; set; }
         public Order Order { get; set; } = null!;
@@ -12,6 +12,6 @@
         public int ProductId { get; set; }
         public Product Product { get; set; } = null!;
 
-        public double GetTotalPrice() => PricePerUnit * Units;
+        public decimal GetTotalPrice() => PricePerUnit * Units;
     }
 }
